requirements.txt
click>=8.1.7
playwright>=1.52.0
beautifulsoup4>=4.12.0
pyyaml>=6.0.0
openai>=1.40.0
